requirements.txt
mysql-connector-python==9.6.0
mysqlconnect==1.0.0
